+++ src/entity/mod.rs
@@ -1,4 +1,5 @@
 use std::any::{Any, TypeId};
+use std::cell::RefCell;
 use std::collections::HashMap;
 
 use error::NotFound;
@@ -35,7 +36,10 @@ impl ComponentBox {
 }
 
 /// The entity builder is used to create an entity with components.
-pub struct EntityBuilder<'a, T> where T: EntityContainer + 'a {
+pub struct EntityBuilder<'a, T>
+where
+    T: EntityContainer + 'a,
+{
     /// The created entity.
     pub entity: Entity,
 
@@ -46,11 +50,21 @@ pub struct EntityBuilder<'a, T> where T: EntityContainer + 'a {
     pub entity_container: &'a mut T,
 }
 
-impl<'a, T> EntityBuilder<'a, T> where T: EntityContainer {
+impl<'a, T> EntityBuilder<'a, T>
+where
+    T: EntityContainer,
+{
     /// Adds a component of type `C` to the entity.
     pub fn with<C: Component>(self, component: C) -> Self {
         self.entity_component_manager
             .register_component(self.entity, component);
+        self
+    }
+
+    /// Adds an entity as `source` for a shared component of type `C`.
+    pub fn with_shared<C: Component>(self, source: Entity) -> Self {
+        self.entity_component_manager
+            .register_shared_component::<C>(self.entity, source);
         self
     }
 
@@ -73,6 +87,8 @@ impl<'a, T> EntityBuilder<'a, T> where T: EntityContainer {
 pub struct EntityComponentManager {
     /// The entities with its components.
     pub entities: HashMap<Entity, HashMap<TypeId, Box<Any>>>,
+
+    pub shared: HashMap<Entity, RefCell<HashMap<TypeId, Entity>>>,
 }
 
 impl EntityComponentManager {
@@ -99,6 +115,16 @@ impl EntityComponentManager {
             .insert(TypeId::of::<C>(), Box::new(component));
     }
 
+    pub fn register_shared_component<C: Component>(&mut self, target: Entity, source: Entity) {
+        if !self.shared.contains_key(&target) {
+            self.shared.insert(target, RefCell::new(HashMap::new()));
+        }
+
+        self.shared[&target]
+            .borrow_mut()
+            .insert(TypeId::of::<C>(), source);
+    }
+
     /// Register a `component_box` for the given `entity`.
     pub fn register_component_box(&mut self, entity: Entity, component_box: ComponentBox) {
         let (type_id, component) = component_box.consume();
@@ -109,20 +135,51 @@ impl EntityComponentManager {
             .insert(type_id, component);
     }
 
-    /// Returns a refernce of a component of type `C` from the given `entity`. If the entity does
-    /// not exists or it dosen't have a component of type `C` `NotFound` will be returned.
-    pub fn borrow_component<C: Component>(&self, entity: Entity) -> Result<&C, NotFound> {
-        self.entities
+    // Search the the target entity in the entity map.
+    fn target_entity_from_shared<C: Component>(&self, entity: Entity) -> Result<Entity, NotFound> {
+        self.shared
             .get(&entity)
             .ok_or_else(|| NotFound::Entity(entity))
             .and_then(|en| {
-                en.get(&TypeId::of::<C>())
-                    .map(|component| {
-                        component.downcast_ref().expect(
-                            "EntityComponentManager.borrow_component: internal downcast error",
-                        )
-                    }).ok_or_else(|| NotFound::Component(TypeId::of::<C>()))
+                en.borrow()
+                    .get(&TypeId::of::<C>())
+                    .map(|entity| *entity)
+                    .ok_or_else(|| NotFound::Component(TypeId::of::<C>()))
             })
+    }
+
+    // Returns the target entity. First search in entities map. If not found search in shared entity map.
+    fn target_entity<C: Component>(&self, entity: Entity) -> Result<Entity, NotFound> {
+        if !self.entities.contains_key(&entity)
+            || !self.entities[&entity].contains_key(&TypeId::of::<C>())
+        {
+            return self.target_entity_from_shared::<C>(entity);
+        }
+
+        Result::Ok(entity)
+    }
+
+    /// Returns a refernce of a component of type `C` from the given `entity`. If the entity does
+    /// not exists or it dosen't have a component of type `C` `NotFound` will be returned.
+    pub fn borrow_component<C: Component>(&self, entity: Entity) -> Result<&C, NotFound> {
+        let target_entity = self.target_entity::<C>(entity);
+
+        match target_entity {
+            Ok(entity) => self
+                .entities
+                .get(&entity)
+                .ok_or_else(|| NotFound::Entity(entity))
+                .and_then(|en| {
+                    en.get(&TypeId::of::<C>())
+                        .map(|component| {
+                            component.downcast_ref().expect(
+                                "EntityComponentManager.borrow_component: internal downcast error",
+                            )
+                        })
+                        .ok_or_else(|| NotFound::Component(TypeId::of::<C>()))
+                }),
+            Err(_) => Result::Err(NotFound::Entity(entity)),
+        }
     }
 
     /// Returns a mutable refernce of a component of type `C` from the given `entity`. If the entity does
@@ -131,22 +188,29 @@ impl EntityComponentManager {
         &mut self,
         entity: Entity,
     ) -> Result<&mut C, NotFound> {
-        self.entities
-            .get_mut(&entity)
-            .ok_or_else(|| NotFound::Entity(entity))
-            .and_then(|en| {
-                en.get_mut(&TypeId::of::<C>())
-                    .map(|component| {
-                        component.downcast_mut().expect(
+        let target_entity = self.target_entity::<C>(entity);
+
+        match target_entity {
+            Ok(entity) => self
+                .entities
+                .get_mut(&entity)
+                .ok_or_else(|| NotFound::Entity(entity))
+                .and_then(|en| {
+                    en.get_mut(&TypeId::of::<C>())
+                        .map(|component| {
+                            component.downcast_mut().expect(
                             "EntityComponentManager.borrow_mut_component: internal downcast error",
                         )
-                    }).ok_or_else(|| NotFound::Component(TypeId::of::<C>()))
-            })
+                        })
+                        .ok_or_else(|| NotFound::Component(TypeId::of::<C>()))
+                }),
+            Err(_) => Result::Err(NotFound::Entity(entity)),
+        }
     }
 }
 
-/// This trait is used to define a custom container for entities. 
-/// A entity container is used for entiy iteration inside of the 
+/// This trait is used to define a custom container for entities.
+/// A entity container is used for entiy iteration inside of the
 /// system's run methods.
 pub trait EntityContainer {
     /// Registers the give 'entity'.
@@ -159,7 +223,7 @@ pub trait EntityContainer {
 /// VecEntityContainer is the default vector based implementation of an entiy container.
 #[derive(Default)]
 pub struct VecEntityContainer {
-    pub inner: Vec<Entity>
+    pub inner: Vec<Entity>,
 }
 
 impl EntityContainer for VecEntityContainer {
@@ -168,8 +232,9 @@ impl EntityContainer for VecEntityContainer {
     }
 
     fn remove_entity(&mut self, entity: Entity) {
-        self.inner.iter()
-        .position(|&n| n == entity)
-        .map(|e| self.inner.remove(e));
+        self.inner
+            .iter()
+            .position(|&n| n == entity)
+            .map(|e| self.inner.remove(e));
     }
 }