+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.2 Shared components (wip)
+
+* Share components between entities
+
 # 0.1.1 Initial release
 
 * Register entities with components using builder